requirements.txt
alembic==1.12.0
annotated-types==0.6.0
anyio==3.7.1
asyncpg==0.28.0
bcrypt==4.0.1
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
dnspython==2.4.2
email-validator==2.0.0.post2
fastapi==0.104.0
fastapi-users==12.1.2
fastapi-users-db-sqlalchemy==6.0.1
Flask==3.0.0
google-api-core==2.12.0
google-api-python-client==2.104.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
greenlet==3.0.0
h11==0.14.0
httpcore==0.18.0
httplib2==0.22.0
httptools==0.6.1
httpx==0.25.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
makefun==1.15.1
Mako==1.2.4
MarkupSafe==2.1.3
oauthlib==3.2.2
orjson==3.9.9
packaging==23.2
passlib==1.7.4
pluggy==1.3.0
protobuf==4.24.4
psycopg2==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
PyJWT==2.8.0
pyparsing==3.1.1
pytest==7.4.3
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
typing_extensions==4.8.0
ujson==5.8.0
uritemplate==4.1.1
urllib3==2.0.6
uvicorn==0.23.2
watchfiles==0.21.0
websockets==11.0.3
Werkzeug==3.0.0
